转移MainActivity的部分职责到MainPresenter中
